+++ internal/services/hit.go
@@ -47,7 +47,7 @@ func (hs *HitService) Add(existsGuest bool, sessionDb entitydb.Session, advRefer
 		AdvUuid:      advReferer.AdvUuid,
 		GuestUuid:    sessionDb.GuestUuid,
 		IsNewGuest:   existsGuest == false,
-		UserId:       statData.UserId,
+		UserId:       uint32(statData.UserId),
 		IsUserAuth:   statData.UserId > 0,
 		Url:          statData.Url,
 		Url404:       statData.IsError404,
